22.06.2021

+++ app/src/main/java/com/example/stopsmoke/User.kt
@@ -4,6 +4,6 @@ package com.example.stopsmoke
 class User (val id: String = "",
             val name: String = "",
             val email: String = "",
-            val cenaPaczki: Float = 0F,
+            val cenaPaczki: Double = 0.0,
             val iloscPapierosow: Int = 0,
             val dataOstatniego: String = "")
